requirements.txt
streamlit==1.50.0
pytesseract==0.3.13
opencv-python==4.12.0.88
Pillow==11.3.0
pandas==2.3.3
PyYAML==6.0.3
python-dateutil==2.9.0.post0
PyMuPDF==1.24.10
scikit-learn==1.5.2
joblib==1.4.2
altair==5.5.0
numpy==2.2.6
flask
gunicorn
requests
python-dotenv
